further test fixes and moving from stable to unstable
